% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateUtilities.R
\name{AddArguments}
\alias{AddArguments}
\title{Add arguments to function evaluation}
\usage{
AddArguments(inputFunction, ...)
}
\arguments{
\item{inputFunction}{A function.}

\item{...}{Additional arguments to be passed to inherited methods}
}
\description{
Add arguments to function evaluation
}
