% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildDT.R
\name{BuildDT}
\alias{BuildDT}
\title{Load all modules}
\usage{
BuildDT(data, title = "Table", fileName, menuOutputOptions)
}
\arguments{
\item{data}{Some data.}

\item{title}{The title of the table.}

\item{fileName}{The name of file to download.}

\item{menuOutputOptions}{Choices for the menu output.}
}
\description{
Load all modules
}
