% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPlot.R
\name{BuildPlot}
\alias{BuildPlot}
\title{Load all modules}
\usage{
BuildPlot(table, title = "Table", fileName, menuOutputOptions)
}
\arguments{
\item{table}{Some data.}

\item{title}{The title of the table.}

\item{fileName}{The name of file to download.}

\item{menuOutputOptions}{Choices for the menu output.}
}
\description{
Load all modules
}
