% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateUtilities.R
\name{DetectChange}
\alias{DetectChange}
\title{Update the function evaluation when something triggers}
\usage{
DetectChange(method, ...)
}
\arguments{
\item{method}{Specify the method for the function.}

\item{...}{Additional arguments to be passed to inherited methods}
}
\description{
Update the function evaluation when something triggers
}
