% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateUtilities.R
\name{DetectChange_Second}
\alias{DetectChange_Second}
\title{Update the function evaluation when something triggers}
\usage{
DetectChange_Second(...)
}
\arguments{
\item{...}{Additional arguments to be passed to inherited methods}
}
\description{
Update the function evaluation when something triggers
}
