% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateUtilities.R
\name{DetectChange_Time}
\alias{DetectChange_Time}
\title{Update the function evaluation when something triggers}
\usage{
DetectChange_Time(triggerTime = "09:30:14")
}
\arguments{
\item{triggerTime}{Time at which the DetectChange_Time function should trigger.}
}
\description{
Update the function evaluation when something triggers
}
