% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R
\name{Documentation}
\alias{Documentation}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
Documentation(
  table,
  title = "Table",
  fileName,
  menuOutputOptions,
  method,
  inputFunction,
  when,
  session = NULL,
  refreshTime = 1,
  text,
  tabName,
  icon,
  triggerTime = "15:32:30",
  ...
)
}
\arguments{
\item{table}{Some data.}

\item{title}{The title of the table.}

\item{fileName}{The name of file to download.}

\item{menuOutputOptions}{Choices for the menu output.}

\item{method}{Specify the method for the function.}

\item{inputFunction}{A function.}

\item{when}{When should the update occur?}

\item{session}{A shiny session.}

\item{refreshTime}{Refresh time of the app in seconds.}

\item{text}{Text to show for the menu item.}

\item{tabName}{The name of a tab that this menu item will activate. Not compatible with href.}

\item{icon}{An icon tag, created by icon. If NULL, don't display an icon.}

\item{triggerTime}{Time at which the DetectChange_Time function should trigger.}

\item{...}{Additional arguments to be passed to inherited methods}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
