% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MenuItem.R
\name{MenuItem}
\alias{MenuItem}
\title{Cooler Menu Item}
\usage{
MenuItem(text, tabName, icon, ...)
}
\arguments{
\item{text}{Text to show for the menu item.}

\item{tabName}{The name of a tab that this menu item will activate. Not compatible with href.}

\item{icon}{An icon tag, created by icon. If NULL, don't display an icon.}

\item{...}{Additional arguments to be passed to inherited methods}
}
\description{
Cooler Menu Item
}
