% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdateUtilities.R
\name{Update}
\alias{Update}
\title{Update the function evaluation when something triggers}
\usage{
Update(inputFunction, when, session = NULL, refreshTime = 1)
}
\arguments{
\item{inputFunction}{A function.}

\item{when}{When should the update occur?}

\item{session}{A shiny session.}

\item{refreshTime}{Refresh time of the app in seconds.}
}
\description{
Update the function evaluation when something triggers
}
