% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Documentation.R
\name{Documentation}
\alias{Documentation}
\title{This function does nothing. It is intended to inherit is parameters'
documentation.}
\usage{
Documentation(data, title = "Table", fileName, menuOutputOptions, ...)
}
\arguments{
\item{data}{Some data.}

\item{title}{The title of the table.}

\item{fileName}{The name of file to download.}

\item{menuOutputOptions}{Choices for the menu output.}

\item{...}{Additional arguments to be passed to inherited methods}
}
\description{
This function does nothing. It is intended to inherit is parameters'
documentation.
}
